requirements.txt
# Core Dependencies
# HTTP requests
requests==2.31.0

# Logging
loguru==0.7.2

# WebSocket support
websockets==12.0

# Configuration management
pyyaml==6.0.1

# Async HTTP client
aiohttp==3.9.1

# Browser automation
playwright==1.40.0

# Protocol buffer handling
blackboxprotobuf==1.1.0

# JavaScript execution in Python
PyExecJS==1.5.1

# Environment variable management
python-dotenv==1.0.0
